+++ analyzers/tests/SonarAnalyzer.TestFramework/Verification/IssueValidation/CompilationIssues.cs
@@ -25,12 +25,12 @@ namespace SonarAnalyzer.TestFramework.Verification.IssueValidation;
 internal sealed class CompilationIssues
 {
     public string LanguageVersion { get; }
-    private readonly DiagnosticVerifier.FileIssueLocations[] fileIssues;
+    private readonly FileIssueLocations[] fileIssues;
 
     public CompilationIssues(string languageVersion, IEnumerable<DiagnosticVerifier.File> files)
     {
         LanguageVersion = languageVersion;
-        fileIssues = files.Select(x => new DiagnosticVerifier.FileIssueLocations(x.FileName, IssueLocationCollector.GetExpectedIssueLocations(x.Content.Lines))).ToArray();
+        fileIssues = files.Select(x => new FileIssueLocations(x.FileName, IssueLocationCollector.GetExpectedIssueLocations(x.Content.Lines))).ToArray();
     }
 
     public CompilationIssues(string languageVersion, Diagnostic[] diagnostics)
@@ -45,7 +45,7 @@ internal sealed class CompilationIssues
             }
         }
         LanguageVersion = languageVersion;
-        fileIssues = map.Select(x => new DiagnosticVerifier.FileIssueLocations(x.Key, x.Value)).ToArray();
+        fileIssues = map.Select(x => new FileIssueLocations(x.Key, x.Value)).ToArray();
 
         void Add(IssueLocationCollector.IssueLocation issue)
         {
